Fix mission form style issues
