+++ frontend/cypress/e2e/side_window/mission_form/main_form.spec.ts
@@ -29,7 +29,7 @@ context('Side Window > Mission Form > Main Form', () => {
 
     cy.get('label').contains('Administration 2').should('exist')
 
-    cy.clickButton('Supprimer cette unité', { index: 1 })
+    cy.clickButton('Supprimer cette unité')
 
     cy.get('label').contains('Administration 2').should('not.exist')
   })
