+++ frontend/src/features/SideWindow/MissionForm/ActionForm/AirSurveillanceForm.tsx
@@ -15,9 +15,9 @@ import { FLIGHT_GOALS_AS_OPTIONS } from './shared/constants'
 import { FleetSegmentsField } from './shared/FleetSegmentsField'
 import { getTitleDateFromUtcStringDate } from './shared/utils'
 import { useNewWindow } from '../../../../ui/NewWindow'
-import { FieldsetGroup } from '../FieldsetGroup'
-import { FormBody } from '../FormBody'
-import { FormHead } from '../FormHead'
+import { FieldsetGroup } from '../shared/FieldsetGroup'
+import { FormBody } from '../shared/FormBody'
+import { FormHead } from '../shared/FormHead'
 
 import type { MissionActionFormValues } from '../types'
 import type { Promisable } from 'type-fest'
