+++ frontend/src/features/SideWindow/MissionForm/ActionForm/LandControlForm.tsx
@@ -21,10 +21,10 @@ import { SpeciesField } from './shared/SpeciesField'
 import { getTitleDateFromUtcStringDate } from './shared/utils'
 import { VesselField } from './shared/VesselField'
 import { useNewWindow } from '../../../../ui/NewWindow'
-import { FieldGroup } from '../FieldGroup'
-import { FieldsetGroup } from '../FieldsetGroup'
-import { FormBody } from '../FormBody'
-import { FormHead } from '../FormHead'
+import { FieldGroup } from '../shared/FieldGroup'
+import { FieldsetGroup } from '../shared/FieldsetGroup'
+import { FormBody } from '../shared/FormBody'
+import { FormHead } from '../shared/FormHead'
 
 import type { MissionActionFormValues } from '../types'
 import type { Promisable } from 'type-fest'
