+++ frontend/src/features/SideWindow/MissionForm/ActionForm/shared/FormikMultiInfractionPicker/index.tsx
@@ -17,8 +17,8 @@ import { Infraction } from './Infraction'
 import { useGetInfractionsQuery } from '../../../../../../api/infraction'
 import { FrontendError } from '../../../../../../libs/FrontendError'
 import { useNewWindow } from '../../../../../../ui/NewWindow'
-import { FieldsetGroup, FieldsetGroupSpinner } from '../../../FieldsetGroup'
-import { FieldsetGroupSeparator } from '../../../FieldsetGroupSeparator'
+import { FieldsetGroup, FieldsetGroupSpinner } from '../../../shared/FieldsetGroup'
+import { FieldsetGroupSeparator } from '../../../shared/FieldsetGroupSeparator'
 import { INFRACTION_TYPES_AS_OPTIONS, MissionActionInfractionSchema } from '../constants'
 
 import type { MissionAction } from '../../../../../../domain/types/missionAction'
