+++ frontend/src/features/SideWindow/MissionForm/ActionForm/shared/GearsField.tsx
@@ -8,8 +8,8 @@ import { FormikMultiInfractionPicker } from './FormikMultiInfractionPicker'
 import { useGetGearsQuery } from '../../../../../api/gear'
 import { FrontendError } from '../../../../../libs/FrontendError'
 import { useNewWindow } from '../../../../../ui/NewWindow'
-import { FieldGroup } from '../../FieldGroup'
-import { FieldsetGroupSpinner } from '../../FieldsetGroup'
+import { FieldGroup } from '../../shared/FieldGroup'
+import { FieldsetGroupSpinner } from '../../shared/FieldsetGroup'
 
 import type { Gear } from '../../../../../domain/types/Gear'
 import type { MissionAction } from '../../../../../domain/types/missionAction'
