+++ frontend/src/features/SideWindow/MissionForm/ActionForm/shared/SpeciesField.tsx
@@ -9,9 +9,9 @@ import { useGetSpeciesQuery } from '../../../../../api/specy'
 import { BOOLEAN_AS_OPTIONS } from '../../../../../constants'
 import { FrontendError } from '../../../../../libs/FrontendError'
 import { useNewWindow } from '../../../../../ui/NewWindow'
-import { FieldGroup } from '../../FieldGroup'
-import { FieldsetGroupSpinner } from '../../FieldsetGroup'
-import { FieldsetGroupSeparator } from '../../FieldsetGroupSeparator'
+import { FieldGroup } from '../../shared/FieldGroup'
+import { FieldsetGroupSpinner } from '../../shared/FieldsetGroup'
+import { FieldsetGroupSeparator } from '../../shared/FieldsetGroupSeparator'
 
 import type { Specy } from '../../../../../domain/types/specy'
 import type { MissionActionFormValues } from '../../types'
