+++ frontend/src/features/SideWindow/MissionForm/MainForm/FormikMultiZonePicker.tsx
@@ -137,6 +137,7 @@ const ZoneWrapper = styled.div`
 
 const Link = styled.a`
   align-items: center;
+  color: ${p => p.theme.color.slateGray};
   cursor: pointer;
   display: inline-flex;
