+++ frontend/src/features/SideWindow/MissionForm/shared/FieldGroup.tsx
