+++ frontend/src/features/SideWindow/MissionForm/shared/FieldsetGroupSeparator.tsx
