+++ frontend/src/ui/FrontendErrorBoundary.tsx
@@ -1,7 +1,6 @@
 /* eslint-disable no-console */
 
 import { Component } from 'react'
-import styled from 'styled-components'
 
 import { FrontendError } from '../libs/FrontendError'
 
@@ -62,60 +61,5 @@ export class FrontendErrorBoundary extends Component<FrontendErrorBoundaryProps,
     const { hasError, isHandled, message, scope } = this.state
 
     return children
-
-    // For the moment, we do not show any error message
-    // @ts-ignore
-    return (
-      <Wrapper>
-        <p>Une erreur est survenue :</p>
-
-        <code>
-          {JSON.stringify(
-            {
-              isHandled,
-              message,
-              scope
-            },
-            null,
-            2
-          )}
-        </code>
-      </Wrapper>
-    )
   }
 }
-
-const Wrapper = styled.div`
-  position: fixed;
-  box-sizing: border-box;
-  inset: 0px;
-  width: 100vw;
-  height: 100vh;
-  background-color: rgba(0, 0, 0, 0.85);
-  color: rgb(232, 232, 232);
-  padding: 2rem;
-  overflow: auto;
-
-  * {
-    box-sizing: border-box;
-  }
-
-  > p {
-    font-size: 18px;
-    line-height: 1;
-    margin: 0 0 32px 0;
-  }
-
-  > code {
-    background-color: black;
-    color: #00ff00;
-    display: block;
-    font-family: monospace;
-    font-size: 18px;
-    line-height: 1.5;
-    margin-top: 1rem;
-    padding: 1rem;
-    white-space: pre-wrap;
-    overflow: auto;
-  }
-`
